The second argument of ioctl is long not int (under Linux)
All our ioctl wrappers were using an int as second argument, but
the glibc ioctl function expects an unsigned long. This wrong prototype
would cause sign extension of the IOCTL request value, thus we would end
up asking for a non existing request. Luckily glibc internally cuts this
down to 32 bits again so things happened to work, but it is better to
fix this and give the wrappers a proper prototype.

+++ contrib/test/capture-example.c
@@ -53,7 +53,7 @@ static void errno_exit(const char *s)
 	exit(EXIT_FAILURE);
 }
 
-static int xioctl(int fh, int request, void *arg)
+static int xioctl(int fh, unsigned long int request, void *arg)
 {
 	int r;
