+++ utils/v4l2-ctl/cx18-ctl.c
@@ -217,7 +217,7 @@ static char *doread(const char *fn)
 	return buf;
 }
 
-static int doioctl(int fd, int request, void *parm, const char *name)
+static int doioctl(int fd, unsigned long int request, void *parm, const char *name)
 {
 	int retVal;
