+++ utils/v4l2-ctl/ivtv-ctl.c
@@ -293,7 +293,7 @@ static const char *field2s(int val)
 	}
 }
 
-static int doioctl(int fd, int request, void *parm, const char *name)
+static int doioctl(int fd, unsigned long int request, void *parm, const char *name)
 {
 	int retVal;
